+++ Recipe Fetch/Recipe Fetch/ContentView.swift
@@ -7,32 +7,6 @@
 
 import SwiftUI
 
-// https://d3jbb8n5wk0qxi.cloudfront.net/recipes.json
-
-struct Welcome: Codable {
-    let recipes: [Recipe]
-}
-
-// MARK: - Recipe
-struct Recipe: Codable {
-    let cuisine: String
-    let name: String
-    let photoURLLarge: String
-    let photoURLSmall: String
-    let sourceURL: String?
-    let uuid: String
-    let youtubeURL: String
-
-    enum CodingKeys: String, CodingKey {
-        case cuisine, name
-        case photoURLLarge = "photo_url_large"
-        case photoURLSmall = "photo_url_small"
-        case sourceURL = "source_url"
-        case uuid
-        case youtubeURL = "youtube_url"
-    }
-}
-
 struct ContentView: View {
     var body: some View {
         VStack {
